chore: refactor TP manager in a dedicated module
