+++ core/control-plane/lib/control-plane-transfer-provision-lib/src/main/java/org/eclipse/edc/connector/controlplane/transfer/provision/ResourceManifestGeneratorImpl.java
