+++ core/control-plane/lib/control-plane-transfer-provision-lib/src/test/java/org/eclipse/edc/connector/controlplane/transfer/provision/ResourceManifestGeneratorImplTest.java
@@ -14,7 +14,7 @@
 
 package org.eclipse.edc.connector.controlplane.transfer.provision;
 
-import org.eclipse.edc.connector.controlplane.transfer.TestResourceDefinition;
+import org.eclipse.edc.connector.controlplane.transfer.provision.fixtures.TestResourceDefinition;
 import org.eclipse.edc.connector.controlplane.transfer.spi.provision.ConsumerResourceDefinitionGenerator;
 import org.eclipse.edc.connector.controlplane.transfer.spi.provision.ProviderResourceDefinitionGenerator;
 import org.eclipse.edc.connector.controlplane.transfer.spi.types.TransferProcess;
