+++ core/control-plane/lib/control-plane-transfer-provision-lib/src/testFixtures/java/org/eclipse/edc/connector/controlplane/transfer/provision/fixtures/TestResourceDefinition.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.connector.controlplane.transfer;
+package org.eclipse.edc.connector.controlplane.transfer.provision.fixtures;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonTypeName;
