+++ settings.gradle.kts
@@ -62,7 +62,9 @@ include(":core:control-plane:control-plane-core")
 include(":core:control-plane:control-plane-aggregate-services")
 include(":core:control-plane:control-plane-transform")
 include(":core:control-plane:control-plane-transfer")
+include(":core:control-plane:control-plane-transfer-manager")
 include(":core:control-plane:lib:control-plane-policies-lib")
+include(":core:control-plane:lib:control-plane-transfer-provision-lib")
 
 include(":core:data-plane:data-plane-util")
 include(":core:data-plane:data-plane-core")
